409 Conflict on Enrollment
A 409 error can indicate that the device record has no name. This happens when the webhook is issued and this program submits only an asset tag. The JSS responds that a name is reqiured. Delaying and trying again seems to work fine. I added a line to sleep for 10 seconds and try again.